requirements.txt
asgiref==3.3.1
attrs==20.3.0
autopep8==1.5.4
Django==3.1.4
django-activity-stream==0.9.0
django-filter==2.4.0
django-reversion==3.0.8
djangorestframework==3.12.2
drf-spectacular==0.11.1
gunicorn==20.0.4
inflection==0.5.1
jsonschema==3.2.0
Markdown==3.3.3
pycodestyle==2.6.0
pyrsistent==0.17.3
pytz==2020.4
PyYAML==5.3.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
